refactor(#267 F3): rename src/say.ts → src/synth.ts, extract voice-routing
The naming-collision audit (#267 P0.F3): `src/say.ts` (public lib API:
`say()`, `buildSayArgs`, `SayError`) and `src/cli/say.ts` (citty CLI
wrapper) share a name in different folders, and `pickVoiceForLang`
(pure routing logic) lived inside the CLI file.

- `src/say.ts` → `src/synth.ts` (the lib half — what `say()` actually
  does). Public surface unchanged: `src/lib.ts` re-exports from `./synth`.
- `pickVoiceForLang` + `RU_DARWIN_FALLBACK_VOICE` → `src/voice-routing.ts`.
  `src/cli/say.ts` imports it; `src/cli.ts` re-exports from the new
  module so the public surface is unchanged.
- `tests/unit/say.test.ts` → split into:
  - `tests/unit/synth.test.ts` (`buildSayArgs`, `--no-expand-abbrev`)
  - `tests/unit/voice-routing.test.ts` (`pickVoiceForLang`)

Closes F10 as a side-effect (routing logic no longer lives in a CLI
file). F13 was inaccurate — the table-driven `pickVoiceForLang` tests
already existed in `say.test.ts:93-122`; this commit just relocates
them into their own file.

`bunx tsc --noEmit` clean. `bun test tests/unit` → 122/122 passing.

Refs #267
