+++ src/synth.ts
