+++ tests/unit/synth.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from "bun:test";
-import { buildSayArgs } from "../../src/say";
-import { pickVoiceForLang } from "../../src/cli";
+import { buildSayArgs } from "../../src/synth";
 
 describe("buildSayArgs", () => {
   it("starts with the 'say' subcommand", () => {
@@ -90,33 +89,3 @@ describe("--no-expand-abbrev (#232)", () => {
   });
 });
 
-describe("pickVoiceForLang (auto-routing)", () => {
-  it("returns en-am_michael for English with high confidence", () => {
-    expect(pickVoiceForLang("en", 0.95)).toBe("en-am_michael");
-  });
-
-  it("returns Milena for Russian on darwin (zero-install AVSpeech path)", () => {
-    expect(pickVoiceForLang("ru", 0.95, "darwin")).toBe(
-      "macos-com.apple.voice.compact.ru-RU.Milena",
-    );
-  });
-
-  it("falls back to ru-vosk-m02 for Russian on non-darwin (Vosk replaces Piper-ruslan, #213)", () => {
-    expect(pickVoiceForLang("ru", 0.95, "linux")).toBe("ru-vosk-m02");
-    expect(pickVoiceForLang("ru", 0.95, "win32")).toBe("ru-vosk-m02");
-  });
-
-  it("returns undefined below 0.5 confidence (too ambiguous)", () => {
-    expect(pickVoiceForLang("ru", 0.3)).toBeUndefined();
-  });
-
-  it("returns undefined for unsupported languages", () => {
-    expect(pickVoiceForLang("fr", 0.95)).toBeUndefined();
-    expect(pickVoiceForLang("de", 0.95)).toBeUndefined();
-  });
-
-  it("returns undefined when code is missing", () => {
-    expect(pickVoiceForLang(undefined, 0.95)).toBeUndefined();
-    expect(pickVoiceForLang("", 0.95)).toBeUndefined();
-  });
-});
